refactor: 검색 기능 리팩토링
- 검색 커서 추상 클래스로 추출(해시태그 기반, 제목 기반)
- 서비스 계층 메서드 추출

+++ backend/bottari/src/main/java/com/bottari/bottaritemplate/repository/BottariTemplateRepository.java
@@ -30,23 +30,23 @@ public interface BottariTemplateRepository extends JpaRepository<BottariTemplate
     List<BottariTemplate> findAllWithMember(final String query);
 
     @Query(value = """
-            SELECT STRAIGHT_JOIN
-                       bt.id AS bottariTemplateId,
-                       bt.title AS title,
-                       bt.taken_count AS takenCount,
-                       bt.created_at AS bottariTemplateCreatedAt,
-                       m.id AS memberId,
-                       m.name AS memberName
-            FROM bottari_template bt
-            JOIN member m ON m.id = bt.member_id 
-            WHERE (:query = '' OR MATCH(bt.title) AGAINST(:query IN BOOLEAN MODE))
-                AND(
-                        bt.created_at < :lastCreatedAt
-                            OR (bt.created_at = :lastCreatedAt AND bt.id < :lastId)
-                        )
-            ORDER BY bt.created_at DESC, bt.id DESC 
-            LIMIT :limit
-        """ ,nativeQuery = true)
+                SELECT STRAIGHT_JOIN
+                           bt.id AS bottariTemplateId,
+                           bt.title AS title,
+                           bt.taken_count AS takenCount,
+                           bt.created_at AS bottariTemplateCreatedAt,
+                           m.id AS memberId,
+                           m.name AS memberName
+                FROM bottari_template bt
+                JOIN member m ON m.id = bt.member_id 
+                WHERE (:query = '' OR MATCH(bt.title) AGAINST(:query IN BOOLEAN MODE))
+                    AND(
+                            bt.created_at < :lastCreatedAt
+                                OR (bt.created_at = :lastCreatedAt AND bt.id < :lastId)
+                            )
+                ORDER BY bt.created_at DESC, bt.id DESC 
+                LIMIT :limit
+            """, nativeQuery = true)
     List<BottariTemplateProjection> findNextByCreatedAt(
             final String query,
             final LocalDateTime lastCreatedAt,
@@ -71,10 +71,10 @@ public interface BottariTemplateRepository extends JpaRepository<BottariTemplate
                     )
             ORDER BY bt.taken_count DESC, bt.id DESC
             LIMIT :limit
-            """ ,nativeQuery = true)
+            """, nativeQuery = true)
     List<BottariTemplateProjection> findNextByTakenCount(
             final String query,
-            final Long lastTakenCount,
+            final Integer lastTakenCount,
             final Long lastId,
             final int limit
     );
